requirements.txt
albumentations==1.2.1
asgiref==3.5.2
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
charset-normalizer==2.1.1
cycler==0.11.0
Django==4.1
efficientnet-pytorch==0.7.1
fonttools==4.37.1
idna==3.3
imageio==2.21.2
imgaug==0.4.0
joblib==1.1.0
kiwisolver==1.4.4
matplotlib==3.5.3
munch==2.5.0
networkx==2.8.6
numpy==1.23.2
onnx==1.12.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
Pillow==9.2.0
pretrainedmodels==0.7.4
protobuf==3.20.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
requests==2.28.1
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
segmentation-models-pytorch==0.3.0
Shapely==1.8.4
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tifffile==2022.8.12
timm==0.4.12
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
